refactor: align fee settings ABI with storage

+++ src/interfaces/IFeeHandler.sol
@@ -12,16 +12,14 @@ interface IFeeHandler {
                            TYPE DECLARATIONS
     //////////////////////////////////////////////////////////////*/
     /// @notice The four parameters that interpolate a purchase fee between `maxFeeRate` and `minFeeRate`.
-    /// @dev Rates are capped at 5% (`MAX_FEE_RATE_CAP`) so they fit uint16. The ABI keeps uint128
-    ///      bounds while FeeHandler checked-casts them to uint112 storage: two bounds plus both rates
-    ///      then occupy one word, and uint112 remains wider than a schedule's uint96 purchase amount.
-    ///      `setFeeRateParams` likewise takes uint256 and checked-casts at the write. Either way
-    ///      `_validateFeeSettings` runs on the values first.
+    /// @dev Two uint112 bounds plus two uint16 rates occupy one storage word. The bounds remain wider
+    ///      than a schedule's uint96 purchase amount. `setFeeRateParams` accepts uint256 values for
+    ///      owner ergonomics and checked-casts them at the write; `_validateFeeSettings` runs first.
     struct FeeSettings {
         uint16 minFeeRate; // the lowest possible fee
         uint16 maxFeeRate; // the highest possible fee
-        uint128 feePurchaseLowerBound; // the purchase amount below which max fee is applied
-        uint128 feePurchaseUpperBound; // the purchase amount above which min fee is applied
+        uint112 feePurchaseLowerBound; // the purchase amount below which max fee is applied
+        uint112 feePurchaseUpperBound; // the purchase amount above which min fee is applied
     }
 
     /*//////////////////////////////////////////////////////////////
